fixed no images crash

+++ data_prep/vegetation/geoscape/03_mosaic_and_transform_images.py
@@ -19,17 +19,17 @@ if platform.system() == "Darwin":
     debug = True
 
     input_list = [{"name": "2m land cover",
-               "input_path": "/Users/s57405/Downloads/SurfaceCover_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Surface Cover/Surface Cover 2M JUNE 2021/Standard",
-               "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_2m_land_cover.tif",
-               "s3_file_path": "geoscape/geoscape_2m_land_cover.tif"},
-              {"name": "30m land cover",
-               "input_path": "/Users/s57405/Downloads/SurfaceCover_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Surface Cover/Surface Cover 30M JUNE 2021/Standard",
-               "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_30m_land_cover.tif",
-               "s3_file_path": "geoscape/geoscape_30m_land_cover.tif"},
-              {"name": "trees",
-               "input_path": "/Users/s57405/Downloads/Trees_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Trees/Trees JUNE 2021/Standard",
-               "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_trees.tif",
-               "s3_file_path": "geoscape/geoscape_trees.tif"}]
+                   "input_path": "/Users/s57405/Downloads/SurfaceCover_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Surface Cover/Surface Cover 2M JUNE 2021/Standard",
+                   "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_2m_land_cover.tif",
+                   "s3_file_path": "geoscape/geoscape_2m_land_cover.tif"},
+                  {"name": "30m land cover",
+                   "input_path": "/Users/s57405/Downloads/SurfaceCover_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Surface Cover/Surface Cover 30M JUNE 2021/Standard",
+                   "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_30m_land_cover.tif",
+                   "s3_file_path": "geoscape/geoscape_30m_land_cover.tif"},
+                  {"name": "trees",
+                   "input_path": "/Users/s57405/Downloads/Trees_JUN21_ALLSTATES_GDA94_GEOTIFF_161/Trees/Trees JUNE 2021/Standard",
+                   "output_file": "/Users/s57405/tmp/bushfire/veg/geoscape_trees.tif",
+                   "s3_file_path": "geoscape/geoscape_trees.tif"}]
 else:
     debug = False
 
@@ -47,7 +47,7 @@ else:
                    "s3_file_path": "geoscape/geoscape_trees.tif"}]
 
 if debug:
-    mga_zones = range(49, 51)
+    mga_zones = range(49, 50)
 else:
     mga_zones = range(49, 57)
 
@@ -64,28 +64,36 @@ for input_dict in input_list:
         start_time = datetime.now()
 
         files_to_mosaic = glob.glob(os.path.join(input_dict["input_path"], f"*_Z{zone}_*.tif"))
-        interim_file = os.path.join(input_dict["input_path"], f"temp_Z{zone}.tif")
 
-        gd = gdal.Warp(interim_file, files_to_mosaic, format="GTiff", options="-r cubic -multi -wm 80% -t_srs EPSG:4326 -co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS -overwrite")
-        del gd
-        warped_files.append(interim_file)
+        if len(files_to_mosaic) > 0:
+            interim_file = os.path.join(input_dict["input_path"], f"temp_Z{zone}.tif")
 
-        print(f"\t- zone {zone} done : {datetime.now() - start_time}")
+            gd = gdal.Warp(interim_file, files_to_mosaic, format="GTiff", options="-r cubic -multi -wm 80% -t_srs EPSG:4326 -co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS -overwrite")
+            del gd
+            warped_files.append(interim_file)
+
+            print(f"\t- zone {zone} done : {datetime.now() - start_time}")
+        else:
+            print(f"\t- zone {zone} has no images : {datetime.now() - start_time}")
 
     # mosaic all merged files and output as a single Cloud Optimised GeoTIFF (COG) for all of AU
     start_time = datetime.now()
     print(f"Processing AU")
 
-    vrt_file = os.path.join(input_dict["input_path"], "temp_au.vrt")
-    my_vrt = gdal.BuildVRT(vrt_file, warped_files)
-    my_vrt = None
+    if len(warped_files) > 0:
+        vrt_file = os.path.join(input_dict["input_path"], "temp_au.vrt")
+        my_vrt = gdal.BuildVRT(vrt_file, warped_files)
+        my_vrt = None
 
-    gd = gdal.Translate(input_dict["output_file"], vrt_file, format="COG", options="-co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS")
-    # gd = gdal.Warp(input_dict["output_file"], warped_files, format="COG", options="-multi -wm 80% -co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS -overwrite")
-    del gd
-    os.remove(vrt_file)
+        gd = gdal.Translate(input_dict["output_file"], vrt_file, format="COG", options="-co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS")
+        # gd = gdal.Warp(input_dict["output_file"], warped_files, format="COG", options="-multi -wm 80% -co BIGTIFF=YES -co COMPRESS=DEFLATE -co NUM_THREADS=ALL_CPUS -overwrite")
+        del gd
+        os.remove(vrt_file)
 
-    print(f"\t- done : {datetime.now() - start_time}")
+        print(f"\t- done : {datetime.now() - start_time}")
+    else:
+        print(f"\t- no files to merge : {datetime.now() - start_time}")
+
     start_time = datetime.now()
 
     # delete interim files
